made service account component as generic and other api updation
